refactor: dir arch
